Added session verification functionality in TOTP access registration form.
Now the user's session is verified when submitting the form. If verification fails, user is redirected to go through email confirmation again. And register TOTP access page always generates a new authenticator key and shows this key uri as the QR code. This helps if the key is exposed to client side once but registration was not completed for some reason, the authenticator key would always be reset on the next registration attempt.

Additionally, the following changes are included in this commit:

1. Added new generic TOTP token provider to project;
2. Added internal RFC6238AuthenticationService using the file below as reference.
   Ref: https://github.com/dotnet/aspnetcore/blob/main/src/Identity/Extensions.Core/src/Rfc6238AuthenticationService.cs;
3. Refactored StartUp code using Dependency Injection extensions. Refactored some other files as well;