Updated TOTP challenge authorization handler and authentication information.
Now, the TOTP challenge authorization handler checks if the value of the Authenticaion Time ("auth_time") claim is within the TOTP Authorization expiry duration, set in AccountOptions, when the "totp_authorization_expiry" claim is missing in the authentication cookie. The functionality to remove expired "totp_authorization_expiry" claims and adding a new one during user sign in is removed from the application.

Additionally, updated the Authorization Method Reference ("amr") claim's value to "mfa" in the authentication cookie. Previously, Duende Identity Server used to set this value to "pwd" by default as it was not set during user sign in.