Updated 'RegisterTOTPAccess' page to reject page refresh after TempData expiry.
Now, the page redirects to SignUp root route when trying to refresh the page after the expiry of TempData.

This helps avoid user TOTP Access compromise when a user leaves the screen or browser unattended for more than 3 minutes or stays in that page for more than 3 minutes. Doing a refresh after this time will result in a redirect to the SignUp controller root route. So none can access the page or refresh it to increase the time window to register their own TOTP Access Authenticator.

Note, this only helps avoiding unwanted user TOTP Access compromise via the 'RegisterTOTPAccess' GET action as the GET action increases TOTP Access registration time window.