Enhanced configuration and secret management.
Now secret values are used automatically in the clients configuration class(Config.cs) from the appropriate appsettings.environment.json files. This is done by exposing a configuration section to the static property "StaticConfiguration" in the Startup.cs class.

In the clients configuration class, client properties are now set based on the environment(as set in the appsettings.environment.json files) the application runs in.